Rename Xgit.Plumbing.Util.WorkingTreeOpt.get/2 to a private function inside the new Plumbing module.
